Split config and behavioral tests
