+++ __tests__/config.test.js
@@ -255,183 +255,4 @@ describe('Configuration parsing and behavior', () => {
 
   });
 
-  describe('v1.1.0 behavior', () => {
-
-    test('cycleServers true', () => {
-      const config = {
-        cycleServers: true,
-        servers: [
-          'foo.bar.com',
-          {server: 'bar.baz.gov', port: 666},
-          'a.b.c',
-          'acb.xyz.def.uvw'
-        ]
-      };
-      const cs = new clockSync(config);
-      // 2 shifts -- idx 2
-      cs.shiftServer();
-      cs.shiftServer();
-      expect(cs.currentIndex).toBe(2);
-      expect(cs.currentServer.server).toBe(config.servers[2]);
-      expect(cs.currentServer.port).toBe(cs.client.defaultNtpPort);
-      // 3 more shifts -- idx 1
-      cs.shiftServer();
-      cs.shiftServer();
-      cs.shiftServer();
-      expect(cs.currentIndex).toBe(1);
-      expect(cs.currentServer.server).toBe(config.servers[1].server);
-      expect(cs.currentServer.port).toBe(config.servers[1].port);
-    });
-
-    test('cycleServers false', () => {
-      const config = {
-        cycleServers: false,
-        servers: [
-          'foo.bar.com',
-          {server: 'bar.baz.gov', port: 666},
-          'a.b.c',
-          'acb.xyz.def.uvw'
-        ]
-      };
-      const cs = new clockSync(config);
-      // 2 shifts -- idx 2
-      cs.shiftServer();
-      cs.shiftServer();
-      expect(cs.currentIndex).toBe(2);
-      expect(cs.currentServer.server).toBe(config.servers[2]);
-      expect(cs.currentServer.port).toBe(cs.client.defaultNtpPort);
-      // 3 more shifts -- idx (config.servers.length - 1)
-      var lastIdx = config.servers.length - 1;
-      cs.shiftServer();
-      cs.shiftServer();
-      cs.shiftServer();
-      expect(cs.currentIndex).toBe(lastIdx);
-      expect(cs.currentServer.server).toBe(config.servers[lastIdx]);
-      expect(cs.currentServer.port).toBe(cs.client.defaultNtpPort);
-    });
-
-    test('setOnline true; startOnline false', done => {
-      client.__setOffsetMs(-1000); // ensure mock ntp is slow
-      const config = {
-        startOnline: false
-      };
-      const cs = new clockSync(config);
-      expect(cs.isOnline).toBe(false);
-      expect(cs.delta).toHaveLength(0); // no initial sync
-      expect(cs.tickId).toBeNull();
-
-      // 1st callback for offline delta time
-      function cb(dt) {
-        expect(dt).toBe(0);
-        cs.setOnline(true); // will trigger 1st update to cs.delta array
-        cs.getDelta(cb2); // will trigger 2st update to cs.delta array
-      }
-
-      // 2nd callback for online delta time; fires 'test done'
-      function cb2(dt) {
-        expect(cs.isOnline).toBe(true);
-        expect(cs.delta).toHaveLength(2);
-        expect(cs.tickId).not.toBeNull();
-        expect(dt).toBeLessThanOrEqual(-1000);
-        done();
-      }
-
-      // start chain of callbacks by inspecting the offline delta time
-      cs.getDelta(cb);
-    });
-
-    test('setOnline false; startOnline true', done => {
-      client.__setOffsetMs(-2000); // ensure mock ntp is slow
-      const config = {
-        startOnline: true
-      };
-      const cs = new clockSync(config);
-      expect(cs.isOnline).toBe(true);
-      expect(cs.delta).toHaveLength(1); // 1 initial sync
-      expect(cs.tickId).not.toBeNull();
-
-      // 1st callback for online delta time
-      function cb(dt) {
-        expect(dt).toBeLessThanOrEqual(-2000);
-        cs.setOnline(false);
-        cs.getDelta(cb2);
-      }
-
-      // 2nd callback for offline delta time; fires 'test done'
-      function cb2(dt) {
-        expect(cs.isOnline).toBe(false);
-        expect(cs.delta).toHaveLength(2); // initial sync + 1st getDelta call
-        expect(cs.tickId).toBeNull();
-        expect(dt).toBe(0);
-        done();
-      }
-
-      // start chain of callbacks by inspecting the online delta time
-      cs.getDelta(cb); // will push 1 update into delta
-    });
-
-    test('setOnline false; startOnline false', done => {
-      client.__setOffsetMs(-2000); // ensure mock ntp is slow
-      const config = {
-        startOnline: false
-      };
-      const cs = new clockSync(config);
-      expect(cs.isOnline).toBe(false);
-      expect(cs.delta).toHaveLength(0);
-      expect(cs.tickId).toBeNull();
-
-      // 1st callback for offline delta time
-      function cb(dt) {
-        expect(dt).toBe(0);
-        cs.setOnline(false); // should have no effect
-        cs.getDelta(cb2);
-      }
-
-      // 2nd callback for offline delta time; fires 'test done'
-      function cb2(dt) {
-        expect(cs.isOnline).toBe(false);
-        expect(cs.delta).toHaveLength(0); // should still be empty
-        expect(cs.tickId).toBeNull();
-        expect(dt).toBe(0);
-        done();
-      }
-
-      // start chain of callbacks by inspecting the online delta time
-      cs.getDelta(cb);
-    });
-
-    test('setOnline true; startOnline true', done => {
-      client.__setOffsetMs(-3000); // ensure mock ntp is slow
-      const config = {
-        startOnline: true
-      };
-      const cs = new clockSync(config);
-      expect(cs.isOnline).toBe(true);
-      expect(cs.delta).toHaveLength(1); // initial sync
-      expect(cs.tickId).not.toBeNull();
-      var tid = cs.tickId;
-
-      // 1st callback for online delta time
-      function cb(dt) {
-        expect(dt).toBeLessThanOrEqual(-3000);
-        cs.setOnline(true); // should have no effect
-        cs.getDelta(cb2); // adds 3rd update to delta
-      }
-
-      // 2nd callback for offline delta time; fires 'test done'
-      function cb2(dt) {
-        expect(cs.isOnline).toBe(true);
-        expect(cs.delta).toHaveLength(3);
-        expect(cs.tickId).not.toBeNull();
-        expect(cs.tickId).toBe(tid); // timer id should be the same
-        expect(dt).toBeLessThanOrEqual(-3000);
-        done();
-      }
-
-      // start chain of callbacks by inspecting the online delta time
-      cs.getDelta(cb); // adds 2nd update to delta
-    });
-
-  });
-
 });
